requirements.txt
fbchat-asyncio
attrs